backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database ORM and migrations
sqlalchemy>=2.0.0
alembic>=1.13.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.4.0
email-validator>=2.1.0

# Data processing and file handling
pandas>=2.1.0
openpyxl>=3.1.0

# Database drivers
psycopg2-binary>=2.9.0  # PostgreSQL support
redis>=5.0.0  # Redis for caching

# HTTP client and testing
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Environment variables
python-dotenv>=1.0.0

# Additional utilities
python-dateutil>=2.8.0
typing-extensions>=4.9.0
anyio>=4.2.0
click>=8.1.0
h11>=0.14.0
idna>=3.6
sniffio>=1.3.0
starlette>=0.27.0
